Final LeaderBorad

+++ ExpServer/controllers/expense.js
@@ -1,4 +1,5 @@
 const Expenses = require('../models/expenses')
+const Users = require('../models/users')
 
 
 const addExpense = async(req,res,next) => {
@@ -6,7 +7,9 @@ const addExpense = async(req,res,next) => {
         const expenses = req.body.Expenses
         const description = req.body.Description
         const category = req.body.Category
-        const data = await Expenses.create({amount: expenses, description: description, category: category})
+        console.log(req.user)
+        const data = await Expenses.create({amount: expenses, description: description, category: category, userId: req.user.id})
+        console.log(data)
         res.status(201).json({expenseDetails: data})
     }
     catch(err){
@@ -19,7 +22,8 @@ const addExpense = async(req,res,next) => {
 
 const getExpense = async(req,res,next) => {
     try{
-        const expenses = await Expenses.findAll();
+        const expenses = await Expenses.findAll({where: {userId: req.user.id}});
+        //req.user.getExpenses()
         res.status(200).json({allExpenses: expenses})
     }
     catch (err){
@@ -33,8 +37,13 @@ const getExpense = async(req,res,next) => {
 const deleteExpense = async(req,res,next) => {
     try{
         const uId = req.params.id;
-        await Expenses.destroy({where: {id: uId}})
-        res.sendStatus(200)
+        const response = await Expenses.destroy({where: {id: uId, userId: req.user.id}})
+        if(response === 0){
+            return res.status(404).json({success: false, message: 'Does not belong to user'})
+        }
+        
+        return res.status(200).json({success:true, message:'Deleted Successfully!!'})
+        
     }
     catch(err){
         console.log(err)
